applications/plasmid-viewer/server/requirements.txt
fastapi
uvicorn[standard]
biopython
python-multipart
aiosqlite
